changed Vec3d setter to Vec3i to avoid issues with directional layer

+++ src/main/java/net/rodofire/easierworldcreator/shape/block/layer/AbstractOuterLayer.java
@@ -5,6 +5,7 @@ import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.ChunkPos;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.math.Vec3i;
 import net.minecraft.world.StructureWorldAccess;
 import net.rodofire.easierworldcreator.blockdata.blocklist.BlockListManager;
 import net.rodofire.easierworldcreator.blockdata.blocklist.DividedBlockListManager;
@@ -18,7 +19,7 @@ import java.util.Map;
 public abstract class AbstractOuterLayer extends AbstractRadialLikeLayer {
     protected float maxDistance;
 
-    AbstractOuterLayer(BlockLayerManager blockLayer, Vec3d center, Vec3d direction) {
+    AbstractOuterLayer(BlockLayerManager blockLayer, Vec3d center, Vec3i direction) {
         super(blockLayer, center, direction);
     }
