+++ src/main/java/net/rodofire/easierworldcreator/shape/block/layer/Layer.java
@@ -5,6 +5,7 @@ import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.ChunkPos;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.math.Vec3i;
 import net.minecraft.world.StructureWorldAccess;
 import net.rodofire.easierworldcreator.blockdata.blocklist.BlockListManager;
 import net.rodofire.easierworldcreator.blockdata.blocklist.DividedBlockListManager;
@@ -49,5 +50,5 @@ public interface Layer {
 
     Vec3d getDirectionVector();
 
-    void setDirectionVector(Vec3d directionVector);
+    void setDirectionVector(Vec3i directionVector);
 }
